requirements.txt
gunicorn==20.1.0
bcrypt==3.2.0
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2 
cycler==0.10.0
cymem==2.0.5
dnspython==2.1.0
email-validator==1.1.2
fastai==2.1.10
fastcore==1.3.19
fastprogress==1.0.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
murmurhash==1.0.5
numpy==1.20.2
packaging==20.9
pandas==1.2.3
pathlib==1.0.1
pathy==0.4.0
Pillow==8.2.0
preshed==3.0.5
pycparser==2.20
pydantic==1.7.3
pyparsing==2.4.7
PySide2==5.15.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
shiboken2==5.15.2
six==1.15.0
smart-open==3.0.0
spacy==3.0.5
spacy-legacy==3.0.2
SQLAlchemy==1.3.23
srsly==2.4.0
thinc==8.0.2
threadpoolctl==2.1.0
torch==1.8.1+cpu
torchaudio==0.8.1
torchvision==0.9.1+cpu
tqdm==4.59.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wasabi==0.8.2
Werkzeug==1.0.1
WTForms==2.3.3
